fix: isolate background completion contracts
